pyproject.toml
[tool.poetry]
name = "conda-hooks"
version = "0.2.2"
description = "Keep anaconda environment files up to date"
authors = ["Fabian Köhler <fabian.koehler@protonmail.ch>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/f-koehler/conda-hooks"
repository = "https://github.com/f-koehler/conda-hooks"

[tool.poetry.dependencies]
python = "^3.8"
PyYAML = "^5.3.1"

[tool.poetry.dev-dependencies]
mypy = "^0.910"
flake8 = "^3.9.2"
pyupgrade = "^2.21.2"
isort = "^5.9.2"
invoke = "^1.6.0"
pre-commit = "^2.13.0"
pytest = "^6.2.4"
black = "^21.7b0"
types-PyYAML = "^5.4.3"
pytest-cov = "^2.12.1"
pytest-sugar = "^0.9.4"
rope = "^0.19.0"

[tool.poetry.scripts]
conda_env_store = "conda_hooks.env_store:main"

[tool.autopub]
project-name = "conda-hooks"
git-username = "f-koehler"
git-email = "fabian.koehler@protonmail.ch"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
